+++ stream-processing-infrastructure/src/it/scala/org/mkuthan/streamprocessing/infrastructure/IntegrationTestFixtures.scala
