+++ stream-processing-infrastructure/src/it/scala/org/mkuthan/streamprocessing/infrastructure/bigquery/SCollectionSyntaxTest.scala
