+++ stream-processing-infrastructure/src/it/scala/org/mkuthan/streamprocessing/infrastructure/bigquery/ScioContextSyntaxTest.scala
