+++ stream-processing-infrastructure/src/it/scala/org/mkuthan/streamprocessing/infrastructure/diagnostic/SCollectionSyntaxTest.scala
