+++ stream-processing-infrastructure/src/it/scala/org/mkuthan/streamprocessing/infrastructure/pubsub/ScioContextSyntaxTest.scala
