+++ stream-processing-infrastructure/src/main/scala/org/mkuthan/streamprocessing/infrastructure/common/IoDiagnostic.scala
