+++ stream-processing-infrastructure/src/main/scala/org/mkuthan/streamprocessing/infrastructure/diagnostic/DiagnosticConfiguration.scala
