+++ stream-processing-infrastructure/src/main/scala/org/mkuthan/streamprocessing/infrastructure/dlq/DlqConfiguration.scala
