+++ stream-processing-infrastructure/src/main/scala/org/mkuthan/streamprocessing/infrastructure/storage/StorageBucket.scala
