+++ stream-processing-shared/src/test/scala/org/mkuthan/streamprocessing/shared/scio/SCollectionSyntaxTest.scala
@@ -1,4 +1,4 @@
-package org.mkuthan.streamprocessing.shared.scio.core
+package org.mkuthan.streamprocessing.shared.scio
 
 import com.spotify.scio.testing._
 import com.spotify.scio.testing.SCollectionMatchers
@@ -6,7 +6,6 @@ import com.spotify.scio.testing.SCollectionMatchers
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
 
-import org.mkuthan.streamprocessing.shared.scio.SCollectionSyntax
 import org.mkuthan.streamprocessing.test.scio.TestScioContext
 
 class SCollectionSyntaxTest extends AnyFlatSpec
